refactor(facade): split lib.rs into exports/* and compat/* layers
Reorganise the 1,451-line `src/lib.rs` into a layered facade per
Issue #4362, Stage 1. Confines `#[cfg]` boilerplate to a small set of
boundary files while preserving the public API surface exactly via
`pub use exports::*;` at the crate root.

New layout:
  src/exports.rs                 cfg aggregation point
  src/exports/common.rs          cross-target re-exports
  src/exports/macros.rs          proc-macro re-exports
  src/exports/native.rs          server-side re-exports
  src/exports/wasm.rs            browser-WASM-side re-exports (placeholder)
  src/compat.rs                  wasm-only shim aggregator
  src/compat/apps.rs             reinhardt_apps WASM shim
  src/compat/urls.rs             urls WASM shim
  src/compat/websockets.rs       WebSocketRouter WASM shim
  src/prelude.rs                 prelude module (extracted)

The crate root now retains crate-level attributes, source-file
`pub mod X;` declarations (required to live alongside their target file),
the layered-facade declarations, the `crate::macros` glob namespace used
by macro-generated code, the compat-shim re-exports that preserve
canonical wasm paths (`crate::reinhardt_apps`, `crate::urls`,
`crate::WebSocketRouter`), and the deprecated `define_views!` macro_rules.

Verified:
- cargo check -p reinhardt-web --no-default-features --features full
- cargo check -p reinhardt-web --no-default-features \
    --features pages,client-router --target wasm32-unknown-unknown
- cargo fmt --check
- cargo clippy -p reinhardt-web --lib -- -D warnings (native + wasm)

Fixes #4364
Refs #4362
